+++ src/tarski/grounding/naive/instantiation.py
@@ -1,4 +1,3 @@
-import itertools
 
 from ...syntax.terms import Constant, Variable
 from ..errors import UnableToGroundError